Added a BHoM abstract class, and properties dictionaries
@al-fisher , @EddyDalton , @adecler  Abstract class added for objects,
this is great. Keeping it as light as possible as the intention here is
for every single BHoM object to inherit from it. The IStructuralObject
then has specific structural object properties and methods. One method
in particular is the GetProperties which will return a dictionary of all
properties of the object - to simplify deconstructs as discussed with
@al-fisher