Added IBHoMGeometry interface to GeometryGroup class
Added IBHoMGeometry to GeometryGroup class, and changed the
List<IBHoMGeometry> to Dictionary<string, IBHoMGeometry>. The next part
of the discussion is whether to remove GeometryGroup altogether, and any
BHoMObject with geometry should simply have a Dictionary of geo. Also
for question is whether we have an interface that BHoM objects WITH
geometry inherit from should implement.