ISpatialObject interface added for discussion
Added to Base, and implemented in structural 'Bar'. If an object is
supposed to be a geometric object, we should have an interface that will
control how the object geometry is named, accessed etc. This will become
very important when we need to represent different geometries of an
object for different levels of detail etc. We should discuss object
geometry asap, as there is a potentially huge optimisation if we force
the idea of similar geometries - identical geometries could be
referenced to different objects with a 'location'. This way, if you had
a very complex geometric object (a car engine) for a car park, you would
only store the geometry once, alongside the location (parking spaces)
for each engine (location).