Geometric Object updates FOR DISCUSSION
Added to Base, and implemented in structural 'Bar' an ISpatialObject
interface. If an object is supposed to be a geometric object, we should
have an interface that will control how the object geometry is named,
accessed etc. This will become very important when we need to represent
different geometries of an object for different levels of details etc.
We should discuss object geometry asap, as there is a potentially huge
optimisation if we force the idea of similar geometries - identical
geometries could be referenced to different objects with a 'location'.
This way, if you had a very complex geometric objets (a car engine) for
a car park, you would only store the geometry once, alongside the
lcoation (parking spaces) for each engine (location). I also added
IBHoHGeometry interface to GeometryGroup and upgraded the GeometryGroup
to a dictionary. A Location property is added to ISpatialObject - this
would be a 'global' property to position an element is space - different
geometries within the object would be able to describe themselves
relative to the object location.