Basic Object Updates
Remove the structural grid object - this should be in architecture and
should also be list of curve rather than a line, so that it is flexible
for all use cases.

Add architectural grid object with basic properties - name (inherited)
and list of ICurve. There shouldn't need to be a 'plane' property as a)
it may be a compound grid and b) this should be a GetPlane or
(TryGetPlane) method in the engine (which tests if it is a single line
of ICurve).

Add architectural level object with Elevation property. GetPlane should
be added to the engine.

Add DesignGroup object in the structural.design namespace to hold basic
properties of a group of design elements for group-verification/design.

Add Beam object in the structural namespace with basic properties.

Add Column object in the structural namespace with basic properties.