Now we have 2 fragments: DiffHashFragment and DiffProjFragment.
This is because the Hash only makes sense together with the Project it's been created within.
(You can create differet hashing rules per project)

DiffProjFragment now contains all the info of the project. This will have to be instantiated prior to do a Diffing. It will be assigned to all IBHoMObjects subject to diffing.

DiffHashFragment contains only the Hash and a reference to its related DiffProjFragment.